+++ routes/auth.route.js
@@ -1,8 +1,7 @@
 import express from "express";
 const router = express.Router();
-import { signUp, signIn } from "../controllers/authController.js";
-import { validate } from "../middleware/validateMiddleware.js";
-
+import { signUp, signIn } from "../controllers/auth.controller.js";
+import { validate } from "../middleware/validate.middleware.js";
 
 router
   .post("/signUp", validate("signUp"), signUp) // * sign up users
